% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{siRNASeqEva}
\alias{siRNASeqEva}
\title{A sequence-based method to evaluate the effectiveness of siRNA in human and other models}
\usage{
siRNASeqEva(
  sirna,
  bam.files = NULL,
  gtf_file = NULL,
  gene = NULL,
  genomic_location = NULL,
  tmp_dir = "temp",
  remove_header = 1,
  remove_overhang = 2,
  paired = TRUE,
  num_threads = 1
)
}
\arguments{
\item{sirna}{A data.frame object, with at least two columns of "id","antisense"}

\item{bam.files}{One or multiple bam file. sorted and indexed bam are recommended}

\item{gtf_file}{A gtf file for gene and transcript annotation. "gtf_file" is not need if  "genomic_location" is not null}

\item{gene}{Symbol of interest gene, e.g. E2F1. It is needed if "gtf_file" is not null}

\item{genomic_location}{A string with format like "chr1:1000:2000". "genomic_location" is not needed if "gtf_file" and "gene" is not null}

\item{tmp_dir}{A directory store the temporary outputs}

\item{remove_header}{A numeric value. For siRNA the 1st 5'nucleic acid is usually not needed during alignment}

\item{remove_overhang}{A numeric value. siRNA usually have a 2-nt overhang and they will be removed during alignment}

\item{paired}{Loigic value. The bam file is derived from paired-end RNA-seq}

\item{num_threads}{The number of threads}
}
\value{
data.frame
}
\description{
A sequence-based method to evaluate the effectiveness of siRNA in human and other models
}
\details{
This function can map the input siRNA sequence to RNA-seq reads and evaluate if (1) siRNA match
}
\examples{
#siRNASeqEva(sirna, bam.files, gtf_file="/path/gtf_file.gtf", gene="E2F1")

}
\author{
Guofeng Meng
}
